Expand reddit internal links
Links inside of reddit can be relative so we have to expand
them for use from the EPUB. Reddit seems to always use a specific
format with and a prefixed slash for its links so assume that's
always the case to be able to to a simple search & replace. Not
robust but unlikely to break.

Fixes #1